fix: EXTRA_RSSI Int→Short for system BT settings + collapsible restart banner + scan event line wrap

+++ app/src/main/java/com/jingyu233/bluetoothhook/hook/BluetoothScanHook.kt
@@ -687,7 +687,7 @@ class BluetoothScanHook(
 
                     val intent = Intent(BluetoothDevice.ACTION_FOUND).apply {
                         putExtra(BluetoothDevice.EXTRA_DEVICE, btDevice)
-                        putExtra(BluetoothDevice.EXTRA_RSSI, device.rssi)
+                        putExtra(BluetoothDevice.EXTRA_RSSI, device.rssi.toShort())
                         putExtra(BluetoothDevice.EXTRA_NAME, device.name)
                     }
                     context.sendBroadcast(intent)
